requirements.txt
fastapi[standard]
uvicorn[standard]
sqlalchemy
psycopg2-binary
alembic
python-jose
pydantic-settings
pytest
httpx
sqlmodel
redis
psutil
asyncpg
greenlet>=3.0.0
google-auth
requests
passlib
bcrypt
razorpay
